# NGGIS-Connector-Source-20260918-085605/backup-v8-swagger-20260917-143322/apiDefinition.swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1",
    "title": "National Grid GIS Enhanced V7",
    "description": "Provides read-only enhanced access to structured National Grid GIS PROD data.",
    "x-ms-api-annotation": {
      "status": "Preview"
    },
    "x-ms-keywords": [
      "National Grid",
      "GIS",
      "ArcGIS"
    ],
    "contact": {
      "name": "National Grid",
      "email": "Grant.Lella@nationalgrid.com"
    }
  },
  "host": "127.0.0.1:5008",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/datasets/{dataset}/tables": {
      "get": {
        "tags": [
          "CdpSampleTable"
        ],
        "summary": "List GIS tables",
        "description": "Returns the available GIS tables for the selected dataset.",
        "operationId": "GetTables",
        "consumes": [],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "dataset",
            "in": "path",
            "description": "Name of dataset",
            "required": true,
            "x-ms-summary": "Dataset name",
            "x-ms-url-encoding": "double",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TablesList"
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        },
        "deprecated": false,
        "x-ms-visibility": "internal",
        "x-ms-pageable": {
          "nextLinkName": "@odata.nextLink"
        }
      }
    },
...
    "/gis/lookups/folders": {
      "get": {
        "operationId": "ListFoldersByRoot",
        "summary": "List folders by root",
        "description": "Lists folders from the selected ArcGIS Web Adaptor.",
        "x-ms-visibility": "internal",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Root"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "value": {
                        "type": "string"
                      },
                      "displayName": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        }
      }
    },
    "/gis/lookups/services": {
      "get": {
        "operationId": "ListServicesByFolder",
        "summary": "List datasets by folder",
        "description": "Lists services from the selected root and folder.",
        "x-ms-visibility": "internal",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Root"
          },
          {
            "name": "folder",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Folder"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "value": {
                        "type": "string"
                      },
                      "displayName": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        }
      }
    },
    "/gis/lookups/tables": {
      "get": {
        "operationId": "ListTablesByService",
        "summary": "List tables by dataset",
        "description": "Lists layers and tables from the selected service.",
        "x-ms-visibility": "internal",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "root",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Root"
          },
          {
            "name": "folder",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Folder"
          },
          {
            "name": "service",
            "in": "query",
            "required": true,
            "type": "string",
            "x-ms-summary": "Dataset name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "value": {
                        "type": "string"
                      },
                      "displayName": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        }
      }
    },
    "/gis/lookups/roots": {
      "get": {
        "operationId": "ListRoots",
        "summary": "List roots",
        "description": "Lists ArcGIS Web Adaptors.",
        "x-ms-visibility": "internal",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "value": {
                        "type": "string"
                      },
                      "displayName": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Operation Failed."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Object": {
      "type": "object",
      "properties": {}
    },
...
    "Table": {
      "description": "Represents a table.",
      "type": "object",
      "properties": {
        "Name": {
          "description": "The name of the table. The name is used at runtime.",
          "type": "string"
        },
        "DisplayName": {
          "description": "The display name of the table.",
          "type": "string"
        },
        "DynamicProperties": {
          "description": "Additional table properties provided by the connector to the clients.",
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/Object"
          },
          "readOnly": true
        }
      }
    },
...
    "TablesList": {
      "description": "Represents a list of tables.",
      "type": "object",
      "properties": {
        "value": {
          "description": "List of Tables",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Table"
          }
        }
      }
    },
...
  },
...
}